requirements.txt
python = "^3.11"
fastapi = "^0.108.0"
uvicorn = "^0.25.0"
pandas = "^2.1.4"
shap = "^0.44.1"
lightgbm = "^4.3.0"
python-dotenv = "^1.0.1"
gunicorn = "^21.2.0"
dagshub = "^0.3.20"